+++ contracts/FlatPricing.sol
@@ -32,7 +32,7 @@ contract FlatPricing is PricingStrategy, Ownable {
 
   // Calculate the current price for buy in amount.
   function calculatePrice(uint value, uint weiRaised, uint tokensSold, address msgSender, uint decimals) public constant returns (uint) {
-    uint multiplier = 10 ** decimals;
+    uint multiplier = uint(10) ** decimals;
     return value.mul(multiplier) / oneTokenInWei;
   }
